Fix handling of the LCD rotary encoder
This is a 2-bit rotary encoder with detents; detents are specced to be at signal 0,0.  That is, the encoder knows when it's in a detent as well as exactly between them.

Don't just count greycode transitions and assume we've stepped after passing somewhere between 2 and 4 of them over a time period; we know exactly what time-independent position we're in.  Only step when we pass midway.

This eliminates all the 'dead' stepping where carefully moving one step at a time often does nothing,
as well as overshooting when stepping 'blips' just past a detent before sttling back in.